MGMT-8064: Introduce dry run mode to the installer
This commit introduces a dry run mode to the installer. The mode is
activated when the `DRY_ENABLE=true` environment is set (or when the
`--dry-run` flag is passed to the installer).

The dry run mode disables several destructive actions performed by the
installer. The purpose of the dry run mode is mostly allowing us to
run a lot of installers on the same machine, without causing harm to
that machine, but still communicating with the service. This is useful
when performing load testing to the service.

See the diff for the exact actions that were disabled, but here's a
summary -
- `extractIgnitionToFS` skipped (it's not needed if we're not really
  installing, and multiple installers running on the same machine
  may compete on the paths required for this, and it's also really slow)
- `updateSingleNodeIgnition` skipped (it's not needed if we're not really
  installing, and because extractIgnitionToFS is no longer performed,
  we're missing the file required)
- `generateSshKeyPair` skipped (multiple installers will compete on those
  files, and also it modifies the keys on the host machine, which is
  something we want to avoid)
- `cleanupInstallDevice` skipped (for obvious reasons)
- `checkLocalhostName` skipped (this function forces a random hostname, we
  don't want that)
- `coreos-installer` replaced - instead of calling `coreos-installer` we'll
  now call `dry-installer` - it's up to the installer dry mode user to
  make sure such executable exists in the PATH - this executable needs
  to print progress similarly to the coreos-installer (since these
  printed lines get sent to the service as installation progress)
- `SystemctlAction` - all systemctl actions are now skipped, they're
  usually destructive an unnecessary in dry run mode.
- `Reboot` - skipped for obvious reasons
- `SetBootOrder` - skipped for obvious reasons
- `CreateOpenshiftSshManifest` - skipped, not needed

Other than disabling some actions, I've also added a new CLI flag
to the installer - the user can now pass `--format-disk` flags,
as many as they want, which will cause the installer to format the given
disks. The reason for this change is that up till now, the service asked
the host to format disks by passing it raw `dd ...` commands that were
chained as `bash -c "dd ...; dd ...; podman run...` commands to
installer, which is problematic because it doesn't give us any way
to disable those `dd` calls when we want a dry run mode - they're
outside of the installer's control. Now that we have the new API
that allows the installer to perform the formatting on its own,
we can disable said formatting in dry run mode, when using newer
versions of the service which respect this new API rather than
using `dd`.

Other unrelated changes in this commit -
- `removePullSecret` bug fixed - when the `PullSecretToken` was empty,
  `ReplaceAll` caused some funky looking strings. This censoring is now
  disabled when the pull secret is empty.